+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ### Added
 
+- `sandstone/Scroller` prop `fadeOut` to show fade-out effect
+- `sandstone/VirtualList.VirtualGridList` prop `noAffordance` to remove affordance when scrolling by keys
 - `sandstone/Item` prop `centered`
 - `sandstone/Panels` and `sandstone/Panels.WizardPanel` props `onTransition` and `onWillTransition`
 - `sandstone/Panels.WizardPanel` prop `noAnimation` to suppress view transition animation